Add current balance and exit position features.
Add current balance that is initially based on capital, but changes to
reflect wins and losses.

Add event to WIN bar that resets the martingale counter and updates the
current balance and capital.

Add event to LOSE bar that increments the martingale counter or applies
loss to current balance and capital if clicked when displaying the last
martingale.

Add EXIT POSITION bar that allows you to cut your losses on any martingale
by updating the current balance and capital using the previous bet's total
loss.